requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
jinja2
python-multipart
python-dotenv
python-jose[cryptography]
passlib[bcrypt]